code restructured to fully scalable
