+++ configs/tsconfig.base.json
@@ -10,5 +10,5 @@
     "declarationMap": true,
     "allowSyntheticDefaultImports": true,
   },
-  "files": ["../src/index.ts"]
+  "include": ["../src/**/*.ts"]
 }
